backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sse-starlette==2.1.3
python-multipart==0.0.12

# ML / Data
xgboost==2.1.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.2.1
scipy==1.14.1
shap==0.51.0
fairlearn==0.10.0

# LangGraph / LangChain
langchain==1.2.10
langchain-core==1.2.17
langchain-openai==1.1.10
langgraph==1.0.10
icd9cms==0.2.1

# Utilities
python-dotenv==1.0.1
kagglehub==0.3.4
joblib==1.4.2
